# documentation/cms-openapi.yaml
openapi: 3.0.3
info:
  title: Buddhist Reading Plans API for Content Creators
  description: |
    API for managing Buddhist reading plans and content delivery.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    - Authenticated users: 1000 requests per hour
    - Unauthenticated users: 100 requests per hour
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@buddhist-plans.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.webuddhist.com/v1
    description: Production server
  - url: https://staging-api.webuddhist.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /cms/plans:
    get:
      tags:
        - Plans
      summary: Get all plans with filtering and pagination
      parameters:
        - name: search
          in: query
          description: Search by plan title
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort plans by field
          schema:
            type: string
            enum:
              - title
              - total_days
              - status
            default: title
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - name: skip
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPlans'

    post:
      tags:
        - Plans
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreate'
      responses:
        '201':
          description: Plan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cms/plans/{plan_id}:
    get:
      tags:
        - Plans
      summary: Get plan details
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan details with days listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanWithDays'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Plans
      summary: Update plan metadata
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanUpdate'
      responses:
        '200':
          description: Plan updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Plans
      summary: Delete plan
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Plan deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cms/plans/{plan_id}/status:
    patch:
      tags:
        - Plans
      summary: Update plan status
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanStatusUpdate'
      responses:
        '200':
          description: Plan status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: Invalid status or plan cannot be published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_status:
                  value:
                    error: "invalid_status"
                    message: "Status must be one of: DRAFT, PUBLISHED, UNPUBLISHED"
                cannot_publish:
                  value:
                    error: "cannot_publish"
                    message: "Plan must have at least one day with content to be published"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Day Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DifficultyLevel:
      type: string
      enum:
        - BEGINNER
        - INTERMEDIATE
        - ADVANCED
    ContentType:
      type: string
      enum:
        - TEXT
        - AUDIO
        - VIDEO
        - IMAGE
        - SOURCE_REFERENCE


    # Authentication Schemas
...
    Plan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        image_url:
          type: string
          format: uri
        total_days:
          type: integer
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - UNPUBLISHED
        subscription_count:
          type: integer
      required:
        - id
        - title
        - description
        - total_days
        - status
        - subscription_count
    PlanStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - UNPUBLISHED
      required:
        - status
    PlanCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
        total_days:
          type: integer
          minimum: 1
        image_url:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - difficulty_level
        - total_days
    PlanUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
        total_days:
          type: integer
          minimum: 1
        image_url:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
    PlanWithDays:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/PlanDay'
      required:
        - id
        - title
        - description
        - days
    PlanDay:
      type: object
      properties:
        id:
          type: string
          format: uuid
        day_number:
          type: integer
        title:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
      required:
        - id
        - day_number
        - title
        - tasks
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        content_type:
          $ref: '#/components/schemas/ContentType'
        content:
          type: string
        estimated_time:
          type: integer
          description: Estimated time in minutes
      required:
        - id
        - title
        - content_type
...
    PaginatedPlans:
      type: object
      properties:
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
        skip:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "Insufficient permissions for this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "The requested resource was not found"
...